Move convertStringsToGroups from Group to TestUtil
