+++ src/main/java/reposense/model/Group.java
@@ -52,23 +52,6 @@ public class Group {
     }
 
     /**
-     * Converts all the strings in {@code groups} into {@code Group} objects. Returns null if {@code groups} is null.
-     * @throws IllegalArgumentException if any of the strings are in invalid formats.
-     */
-    public static List<Group> convertStringsToGroups(List<String> groups) throws IllegalArgumentException {
-        if (groups == null) {
-            return null;
-        }
-
-        return groups.stream()
-            .map(temp -> {
-                String[] elements = temp.split(":");
-                Group obj = new Group(elements[0], Arrays.asList(elements[1].split(";")));
-                return obj;
-            }).collect(Collectors.toList());
-    }
-
-    /**
      * Checks that {@code value} is a valid group.
      * @throws IllegalArgumentException if {@code value} does not meet the criteria.
      */
